nat: extract scoping tests into tests_scope.rs (#4409)
Move the 16 interface / routing-instance scoping tests -- the
source_nat_*_scope, static_nat_*_from_interface_scope /
*_source_address_* / *_scope_differentiated_rules_coexist_3605 and
destination_nat_from_{interface,routing_instance}_scope cases that
verify a rule-set scoped to a named interface or VRF matches only its
traffic -- verbatim into tests_scope.rs. No test-local helpers.

Pure code motion: no test body edited, nothing renamed. nat/tests.rs
now holds 16 tests.
