+++ userspace-dp/src/nat/tests.rs
@@ -18,590 +18,6 @@ use crate::{
 use std::collections::BTreeSet;
 use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
 
-// Source NAT scoped by `from interface`: matches a flow ingressing the named
-// interface, NOT another. Fail-on-revert: dropping the `scope_matches` gate in
-// SourceNatRule::matches makes the second assertion (wrong-interface) match and
-// so RED.
-#[test]
-fn source_nat_from_interface_scope_matches_only_named_iface() {
-    let rules = parse_source_nat_rules(&[SourceNATRuleSnapshot {
-        name: "snat".to_string(),
-        from_interface: "ge-0/0/1.0".to_string(),
-        source_addresses: vec!["0.0.0.0/0".to_string()],
-        interface_mode: true,
-        ..SourceNATRuleSnapshot::default()
-    }]);
-    let src: IpAddr = "10.0.61.102".parse().unwrap();
-    let dst: IpAddr = "172.16.80.200".parse().unwrap();
-    let egress_v4 = Some("172.16.80.8".parse().unwrap());
-
-    // Ingress on the named interface -> matches.
-    let on_iface = NatScopeCtx {
-        ingress_ifname: "ge-0/0/1.0",
-        ..NatScopeCtx::default()
-    };
-    assert!(
-        match_source_nat(&rules, &on_iface, "", "", src, dst, egress_v4, None).is_some(),
-        "from-interface-scoped SNAT must match traffic ingressing the named interface"
-    );
-
-    // Ingress on a DIFFERENT interface -> no match.
-    let other_iface = NatScopeCtx {
-        ingress_ifname: "ge-0/0/2.0",
-        ..NatScopeCtx::default()
-    };
-    assert!(
-        match_source_nat(&rules, &other_iface, "", "", src, dst, egress_v4, None).is_none(),
-        "from-interface-scoped SNAT must NOT match traffic ingressing another interface"
-    );
-}
-
-// Source NAT scoped by `from routing-instance`: matches only the named VRF.
-#[test]
-fn source_nat_from_routing_instance_scope_matches_only_named_vrf() {
-    let rules = parse_source_nat_rules(&[SourceNATRuleSnapshot {
-        name: "snat".to_string(),
-        from_routing_instance: "VR1".to_string(),
-        source_addresses: vec!["0.0.0.0/0".to_string()],
-        interface_mode: true,
-        ..SourceNATRuleSnapshot::default()
-    }]);
-    let src: IpAddr = "10.0.61.102".parse().unwrap();
-    let dst: IpAddr = "172.16.80.200".parse().unwrap();
-    let egress_v4 = Some("172.16.80.8".parse().unwrap());
-
-    let in_vr1 = NatScopeCtx {
-        ingress_routing_instance: "VR1",
-        ..NatScopeCtx::default()
-    };
-    assert!(
-        match_source_nat(&rules, &in_vr1, "", "", src, dst, egress_v4, None).is_some(),
-        "from-routing-instance-scoped SNAT must match traffic in the named VRF"
-    );
-
-    let in_vr2 = NatScopeCtx {
-        ingress_routing_instance: "VR2",
-        ..NatScopeCtx::default()
-    };
-    assert!(
-        match_source_nat(&rules, &in_vr2, "", "", src, dst, egress_v4, None).is_none(),
-        "from-routing-instance-scoped SNAT must NOT match traffic in another VRF"
-    );
-}
-
-// Source NAT scoped by `to interface`: matches only the named egress interface.
-#[test]
-fn source_nat_to_interface_scope_matches_only_named_egress() {
-    let rules = parse_source_nat_rules(&[SourceNATRuleSnapshot {
-        name: "snat".to_string(),
-        to_interface: "ge-0/0/2.0".to_string(),
-        source_addresses: vec!["0.0.0.0/0".to_string()],
-        interface_mode: true,
-        ..SourceNATRuleSnapshot::default()
-    }]);
-    let src: IpAddr = "10.0.61.102".parse().unwrap();
-    let dst: IpAddr = "172.16.80.200".parse().unwrap();
-    let egress_v4 = Some("172.16.80.8".parse().unwrap());
-
-    let to_named = NatScopeCtx {
-        egress_ifname: "ge-0/0/2.0",
-        ..NatScopeCtx::default()
-    };
-    assert!(
-        match_source_nat(&rules, &to_named, "", "", src, dst, egress_v4, None).is_some(),
-        "to-interface-scoped SNAT must match traffic egressing the named interface"
-    );
-    let to_other = NatScopeCtx {
-        egress_ifname: "ge-0/0/9.0",
-        ..NatScopeCtx::default()
-    };
-    assert!(
-        match_source_nat(&rules, &to_other, "", "", src, dst, egress_v4, None).is_none(),
-        "to-interface-scoped SNAT must NOT match traffic egressing another interface"
-    );
-}
-
-// A zone-scoped source NAT rule is unaffected by the #3096 scope plumbing
-// (no-regression): empty interface/RI scope = wildcard, so any interface/VRF
-// in the matching zone pair still matches.
-#[test]
-fn source_nat_zone_scope_unaffected_by_interface_plumbing() {
-    let rules = parse_source_nat_rules(&[SourceNATRuleSnapshot {
-        name: "snat".to_string(),
-        from_zone: "lan".to_string(),
-        to_zone: "wan".to_string(),
-        source_addresses: vec!["0.0.0.0/0".to_string()],
-        interface_mode: true,
-        ..SourceNATRuleSnapshot::default()
-    }]);
-    let src: IpAddr = "10.0.61.102".parse().unwrap();
-    let dst: IpAddr = "172.16.80.200".parse().unwrap();
-    let egress_v4 = Some("172.16.80.8".parse().unwrap());
-    // Arbitrary interface names in the scope ctx must not affect a zone rule.
-    let scope = NatScopeCtx {
-        ingress_ifname: "ge-0/0/7.0",
-        egress_ifname: "ge-0/0/8.0",
-        ingress_routing_instance: "whatever",
-        egress_routing_instance: "whatever",
-    };
-    assert!(
-        match_source_nat(&rules, &scope, "lan", "wan", src, dst, egress_v4, None).is_some(),
-        "zone-scoped SNAT must still match regardless of interface/RI ctx"
-    );
-}
-
-// Static NAT (DNAT direction) scoped by `from interface`: the inbound
-// translation fires only when the packet ingresses the named interface.
-// Fail-on-revert: dropping the interface gate in match_dnat_with_counter_scoped
-// makes the wrong-interface assertion fire and so RED.
-#[test]
-fn static_nat_dnat_from_interface_scope_matches_only_named_iface() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            name: "stat".to_string(),
-            from_interface: "ge-0/0/1.0".to_string(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            ..StaticNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let ext: IpAddr = "203.0.113.10".parse().unwrap();
-    // On the named interface -> DNAT applies.
-    assert!(
-        table
-            .match_dnat_with_counter_scoped(ext, 0, None, "", "ge-0/0/1.0", "")
-            .is_some(),
-        "interface-scoped static DNAT must match on the named ingress interface"
-    );
-    // On a different interface -> no DNAT.
-    assert!(
-        table
-            .match_dnat_with_counter_scoped(ext, 0, None, "", "ge-0/0/2.0", "")
-            .is_none(),
-        "interface-scoped static DNAT must NOT match on another ingress interface"
-    );
-}
-
-// Static NAT (SNAT/reverse direction) scoped by `from interface`: the reverse
-// source translation fires only when the packet egresses the named interface.
-#[test]
-fn static_nat_snat_from_interface_scope_matches_only_named_egress() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            source_addresses: Vec::new(),
-            name: "stat".to_string(),
-            from_interface: "ge-0/0/1.0".to_string(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            ..StaticNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let int: IpAddr = "192.168.1.10".parse().unwrap();
-    assert!(
-        table
-            .match_snat_with_counter_scoped(int, 0, None, "", "ge-0/0/1.0", "")
-            .is_some(),
-        "interface-scoped static SNAT must match on the named egress interface"
-    );
-    assert!(
-        table
-            .match_snat_with_counter_scoped(int, 0, None, "", "ge-0/0/2.0", "")
-            .is_none(),
-        "interface-scoped static SNAT must NOT match on another egress interface"
-    );
-}
-
-// #3435 (H01): static-NAT `match source-address` must gate the inbound DNAT
-// translation on the packet SOURCE. Before #3435 the constraint was dropped at
-// the snapshot boundary, installing an all-source mapping that exposed the
-// internal host to ANY source. Fail-on-revert: dropping the `source_ok` gate
-// makes the out-of-scope-source assertion (which expects None) match and so RED.
-#[test]
-fn static_nat_dnat_source_address_gates_inbound() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            name: "stat".to_string(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            source_addresses: vec!["198.51.100.0/24".to_string()],
-            ..StaticNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let ext: IpAddr = "203.0.113.10".parse().unwrap();
-    let in_scope: IpAddr = "198.51.100.7".parse().unwrap();
-    let out_scope: IpAddr = "203.0.113.99".parse().unwrap();
-    assert!(
-        table
-            .match_dnat_with_counter_scoped(ext, 0, Some(in_scope), "", "", "")
-            .is_some(),
-        "static DNAT must fire for a source inside `match source-address`"
-    );
-    assert!(
-        table
-            .match_dnat_with_counter_scoped(ext, 0, Some(out_scope), "", "", "")
-            .is_none(),
-        "static DNAT must NOT fire for a source outside `match source-address` (the #3435 fail-open)"
-    );
-}
-
-// #3435 (M02): a bracket / repeated source-address list must honor EVERY prefix
-// (the multi-value scalar loss). Fail-on-revert: if the compiler/snapshot drops
-// every entry after the first, the second-prefix source goes unmatched and the
-// loop assertion RED.
-#[test]
-fn static_nat_dnat_source_address_bracket_list() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            name: "stat".to_string(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            source_addresses: vec![
-                "198.51.100.0/24".to_string(),
-                "203.0.113.200".to_string(),
-            ],
-            ..StaticNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let ext: IpAddr = "203.0.113.10".parse().unwrap();
-    for src in ["198.51.100.7", "203.0.113.200"] {
-        let s: IpAddr = src.parse().unwrap();
-        assert!(
-            table
-                .match_dnat_with_counter_scoped(ext, 0, Some(s), "", "", "")
-                .is_some(),
-            "static DNAT must fire for source {src} in the bracket list"
-        );
-    }
-    let out: IpAddr = "10.0.0.1".parse().unwrap();
-    assert!(
-        table
-            .match_dnat_with_counter_scoped(ext, 0, Some(out), "", "", "")
-            .is_none(),
-        "static DNAT must NOT fire for a source outside the bracket list"
-    );
-}
-
-// #3435: the reverse (SNAT) direction gates on the packet DESTINATION (the
-// original client), symmetric with the inbound source gate. Fail-on-revert:
-// dropping the source gate lets an out-of-scope destination match.
-#[test]
-fn static_nat_snat_source_address_gates_reverse() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            name: "stat".to_string(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            source_addresses: vec!["198.51.100.0/24".to_string()],
-            ..StaticNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let int: IpAddr = "192.168.1.10".parse().unwrap();
-    let in_scope: IpAddr = "198.51.100.7".parse().unwrap();
-    let out_scope: IpAddr = "203.0.113.99".parse().unwrap();
-    assert!(
-        table
-            .match_snat_with_counter_scoped(int, 0, Some(in_scope), "", "", "")
-            .is_some(),
-        "reverse static SNAT must fire toward a client inside `match source-address`"
-    );
-    assert!(
-        table
-            .match_snat_with_counter_scoped(int, 0, Some(out_scope), "", "", "")
-            .is_none(),
-        "reverse static SNAT must NOT fire toward a client outside `match source-address`"
-    );
-}
-
-// #3435: a source-constrained rule whose source list parses to NOTHING (every
-// entry malformed) must match NO source — fail CLOSED, never revert to
-// match-any (the #2394/#3435 fail-open guard). Fail-on-revert: collapsing the
-// unparseable list to "unconstrained" makes the assertion (expecting None) RED.
-#[test]
-fn static_nat_dnat_source_address_unparseable_fails_closed() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            name: "stat".to_string(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            source_addresses: vec!["not-an-ip".to_string()],
-            ..StaticNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let ext: IpAddr = "203.0.113.10".parse().unwrap();
-    let any: IpAddr = "198.51.100.7".parse().unwrap();
-    assert!(
-        table
-            .match_dnat_with_counter_scoped(ext, 0, Some(any), "", "", "")
-            .is_none(),
-        "a source-scoped rule with no parseable prefix must match nothing (fail closed)"
-    );
-}
-
-// #3435: guard against over-gating — an UNscoped static-NAT rule (no `match
-// source-address`) must still translate every source.
-#[test]
-fn static_nat_dnat_no_source_constraint_matches_any() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[StaticNATRuleSnapshot {
-            name: "stat".to_string(),
-            external_ip: "203.0.113.10".to_string(),
-            internal_ip: "192.168.1.10".to_string(),
-            ..StaticNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let ext: IpAddr = "203.0.113.10".parse().unwrap();
-    let any: IpAddr = "10.9.8.7".parse().unwrap();
-    assert!(
-        table
-            .match_dnat_with_counter_scoped(ext, 0, Some(any), "", "", "")
-            .is_some(),
-        "an unscoped static DNAT must match any source"
-    );
-}
-
-// === #3605: zone/scope-differentiated static-NAT rules must COEXIST ===
-
-// Split-horizon static DNAT: two rules share the same `(external_ip,
-// match-port)` key but differ by `from zone`, translating the same public
-// address to a DIFFERENT internal host per ingress zone. Both must be
-// installed and matched by scope.
-//
-// Fail-on-revert: with the pre-#3605 single-entry map the second rule
-// OVERWRITES the first (last-write-wins), so the untrust packet — whose only
-// surviving entry is the dmz-scoped rule — fails `zone_ok` and returns None
-// (a silent UNtranslated leak). The untrust assertion below then RED.
-#[test]
-fn static_nat_dnat_scope_differentiated_rules_coexist_3605() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[
-            StaticNATRuleSnapshot {
-                name: "from-untrust".to_string(),
-                from_zone: "untrust".to_string(),
-                external_ip: "203.0.113.10".to_string(),
-                internal_ip: "192.168.1.10".to_string(),
-                ..StaticNATRuleSnapshot::default()
-            },
-            StaticNATRuleSnapshot {
-                name: "from-dmz".to_string(),
-                from_zone: "dmz".to_string(),
-                external_ip: "203.0.113.10".to_string(),
-                internal_ip: "10.0.0.10".to_string(),
-                ..StaticNATRuleSnapshot::default()
-            },
-        ],
-        &counters,
-    );
-    let ext: IpAddr = "203.0.113.10".parse().unwrap();
-    // Packet ingressing untrust -> the untrust rule (192.168.1.10). This is the
-    // rule OVERWRITTEN pre-#3605, so this assertion is the fail-on-revert.
-    let via_untrust = table
-        .match_dnat_with_counter_scoped(ext, 0, None, "untrust", "", "")
-        .map(|(d, _)| d.rewrite_dst);
-    assert_eq!(
-        via_untrust,
-        Some(Some("192.168.1.10".parse().unwrap())),
-        "untrust-scoped static DNAT must survive alongside the dmz-scoped rule (#3605 overwrite)"
-    );
-    // Packet ingressing dmz -> the dmz rule (10.0.0.10).
-    let via_dmz = table
-        .match_dnat_with_counter_scoped(ext, 0, None, "dmz", "", "")
-        .map(|(d, _)| d.rewrite_dst);
-    assert_eq!(
-        via_dmz,
-        Some(Some("10.0.0.10".parse().unwrap())),
-        "dmz-scoped static DNAT must translate to its own internal host"
-    );
-}
-
-// Reverse (SNAT) direction of the same #3605 overwrite: two rules map DIFFERENT
-// external addresses to the SAME internal host, scoped by different egress
-// zones. The SNAT map keys on `(internal_ip, snat_port)`, so both rules collide
-// on one key and the pre-#3605 map kept only the last -> the trust-egress
-// return packet un-NATs to the wrong (or no) external address.
-//
-// Fail-on-revert: the trust assertion RED because its rule was overwritten by
-// the wan-scoped rule.
-#[test]
-fn static_nat_snat_scope_differentiated_rules_coexist_3605() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[
-            StaticNATRuleSnapshot {
-                name: "egress-trust".to_string(),
-                from_zone: "trust".to_string(),
-                external_ip: "203.0.113.10".to_string(),
-                internal_ip: "192.168.1.10".to_string(),
-                ..StaticNATRuleSnapshot::default()
-            },
-            StaticNATRuleSnapshot {
-                name: "egress-wan".to_string(),
-                from_zone: "wan".to_string(),
-                external_ip: "203.0.113.20".to_string(),
-                internal_ip: "192.168.1.10".to_string(),
-                ..StaticNATRuleSnapshot::default()
-            },
-        ],
-        &counters,
-    );
-    let int: IpAddr = "192.168.1.10".parse().unwrap();
-    // Return packet egressing trust -> un-NAT to 203.0.113.10 (overwritten rule).
-    let via_trust = table
-        .match_snat_with_counter_scoped(int, 0, None, "trust", "", "")
-        .map(|(d, _)| d.rewrite_src);
-    assert_eq!(
-        via_trust,
-        Some(Some("203.0.113.10".parse().unwrap())),
-        "trust-egress static SNAT must survive alongside the wan-scoped rule (#3605 overwrite)"
-    );
-    // Return packet egressing wan -> un-NAT to 203.0.113.20.
-    let via_wan = table
-        .match_snat_with_counter_scoped(int, 0, None, "wan", "", "")
-        .map(|(d, _)| d.rewrite_src);
-    assert_eq!(
-        via_wan,
-        Some(Some("203.0.113.20".parse().unwrap())),
-        "wan-egress static SNAT must un-NAT to its own external address"
-    );
-}
-
-// #3605 specificity tiering: a zone-SCOPED rule and a zone-WILDCARD rule share
-// the same `(external_ip, match-port)` key. A packet matching the scoped zone
-// must hit the SPECIFIC rule, not the wildcard — regardless of config order
-// (the wildcard is authored FIRST here, the adversarial order). Traffic in any
-// other zone falls through to the wildcard.
-//
-// Fail-on-revert: with the single-entry map the scoped rule (authored second)
-// overwrites the wildcard, so the "other zone" packet — whose only entry is now
-// the trust-scoped rule — returns None instead of the wildcard translation, so
-// the other-zone assertion RED.
-#[test]
-fn static_nat_dnat_scoped_rule_wins_over_coexisting_wildcard_3605() {
-    let counters = NatCounterStore::default();
-    let table = StaticNatTable::from_snapshots(
-        &[
-            StaticNATRuleSnapshot {
-                name: "wildcard".to_string(),
-                from_zone: String::new(), // any zone
-                external_ip: "203.0.113.10".to_string(),
-                internal_ip: "10.0.0.99".to_string(),
-                ..StaticNATRuleSnapshot::default()
-            },
-            StaticNATRuleSnapshot {
-                name: "trust-specific".to_string(),
-                from_zone: "trust".to_string(),
-                external_ip: "203.0.113.10".to_string(),
-                internal_ip: "192.168.1.10".to_string(),
-                ..StaticNATRuleSnapshot::default()
-            },
-        ],
-        &counters,
-    );
-    let ext: IpAddr = "203.0.113.10".parse().unwrap();
-    // trust ingress -> the SPECIFIC rule wins over the wildcard.
-    let via_trust = table
-        .match_dnat_with_counter_scoped(ext, 0, None, "trust", "", "")
-        .map(|(d, _)| d.rewrite_dst);
-    assert_eq!(
-        via_trust,
-        Some(Some("192.168.1.10".parse().unwrap())),
-        "the zone-specific rule must win over a coexisting wildcard regardless of order (#3605)"
-    );
-    // any other zone -> the wildcard catch-all (fail-on-revert: overwritten).
-    let via_other = table
-        .match_dnat_with_counter_scoped(ext, 0, None, "untrust", "", "")
-        .map(|(d, _)| d.rewrite_dst);
-    assert_eq!(
-        via_other,
-        Some(Some("10.0.0.99".parse().unwrap())),
-        "an unmatched zone must fall through to the coexisting wildcard rule (#3605 overwrite)"
-    );
-}
-
-// Destination NAT scoped by `from interface`: the DNAT entry fires only for the
-// named ingress interface. Fail-on-revert: dropping the scope_ok gate in
-// match_entries makes the wrong-interface assertion fire and so RED.
-#[test]
-fn destination_nat_from_interface_scope_matches_only_named_iface() {
-    let counters = NatCounterStore::default();
-    let table = DnatTable::from_snapshots(
-        &[DestinationNATRuleSnapshot {
-            name: "dnat".to_string(),
-            from_interface: "ge-0/0/1.0".to_string(),
-            destination_address: "203.0.113.10".to_string(),
-            pool_address: "10.0.0.5".to_string(),
-            ..DestinationNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let src: IpAddr = "198.51.100.1".parse().unwrap();
-    let dst: IpAddr = "203.0.113.10".parse().unwrap();
-    assert!(
-        table
-            .lookup_with_counter_scoped(PROTO_TCP, src, dst, 0, 443, "", "ge-0/0/1.0", "", None)
-            .is_some(),
-        "interface-scoped DNAT must match on the named ingress interface"
-    );
-    assert!(
-        table
-            .lookup_with_counter_scoped(PROTO_TCP, src, dst, 0, 443, "", "ge-0/0/2.0", "", None)
-            .is_none(),
-        "interface-scoped DNAT must NOT match on another ingress interface"
-    );
-}
-
-// Destination NAT scoped by `from routing-instance`.
-#[test]
-fn destination_nat_from_routing_instance_scope_matches_only_named_vrf() {
-    let counters = NatCounterStore::default();
-    let table = DnatTable::from_snapshots(
-        &[DestinationNATRuleSnapshot {
-            name: "dnat".to_string(),
-            from_routing_instance: "VR1".to_string(),
-            destination_address: "203.0.113.10".to_string(),
-            pool_address: "10.0.0.5".to_string(),
-            ..DestinationNATRuleSnapshot::default()
-        }],
-        &counters,
-    );
-    let src: IpAddr = "198.51.100.1".parse().unwrap();
-    let dst: IpAddr = "203.0.113.10".parse().unwrap();
-    assert!(
-        table
-            .lookup_with_counter_scoped(PROTO_TCP, src, dst, 0, 443, "", "", "VR1", None)
-            .is_some(),
-        "RI-scoped DNAT must match in the named VRF"
-    );
-    assert!(
-        table
-            .lookup_with_counter_scoped(PROTO_TCP, src, dst, 0, 443, "", "", "VR2", None)
-            .is_none(),
-        "RI-scoped DNAT must NOT match in another VRF"
-    );
-}
-
-// === #3429: source-NAT `match destination-port` / `match application`
-// enforcement. Before the fix these match fields were parsed and compiled but
-// never carried to the dataplane, so a port/app-scoped source-NAT rule
-// (including a `then source-nat off` exemption) silently widened to every
-// port/protocol. Reverting the l4_matches gate makes the wrong-port / wrong-app
-// lookups translate, turning the NoMatch assertions RED. ===
-
 #[test]
 fn source_nat_match_destination_port_constrains_flow_3429() {
     let rules = parse_source_nat_rules(&[SourceNATRuleSnapshot {
